Title: 알파벳과 사칙연산, Time: 119ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240419/알파벳과 사칙연산/calculations-with-alphabet.java
@@ -5,7 +5,7 @@ public class Main {
     public static int[] alp = new int[6];
 
     public static String str;
-    public static int ans = 0;
+    public static long ans = 0;
 
     public static void main(String[] args) {
         // 여기에 코드를 작성해주세요.
